Correct now.
I don’t know why it is written in the lecture that “to increase the accuracy of calculations, it is more advisable to use not x0 but xi that stands before the interpolation node”, and at the same time, the polynomial must necessarily coincide with the lagrenge (although it is obvious that using fewer points, we get a less accurate result ), but now everything works so that I get a pass for the lab. I renamed the previous, "more accurate" function and left it anyway.